Add files via upload
Upload setup program first to set the values of gyro and accelerometer into the eeprom.

after the setup is finished, upload the esc calibrate to make sure you have the proper values set and to calibrate the escs to make sure they all start at the same time. Use esc calibrate program to balance the propellers and the vibrations of motors.

after calibrating the escs, motors, and propellers, upload the final program for the drone flight.